#17 / Fixing search & filter
- Moved files and replaced index with test file
- Added search and filter function that finally works !! with checkbox
- Created geo page as an empty page
- Updating airtable API and syncing it to base
