--- test.html
@@ -1,163 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" >
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="style.css"/>
-  <title>Find a Survey from the U.S. Census Bureau</title>
-</head>
-<body>
-        <nav class="navbar navbar-expand-lg navbar-dark">
-            <img src="assets/logo.jpg" width="100" class="mr-5 d-inline-block align-top" alt="U.S. Census Bureau">
-            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-              <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-              <ul class="navbar-nav">
-                <li class="nav-item active">
-                  <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
-                </li>
-                <li class="nav-item">
-                  <a class="nav-link" href="#">Geographies</a>
-                </li>
-                <li class="nav-item">
-                  <a class="nav-link" href="#">Help</a>
-                </li>
-                <li class="nav-item">
-                  <a class="nav-link" href="mailto:amy.lo@census.gov">Feedback</a>
-                </li>
-              </ul>
-            </div>
-        </nav>
-
-        <div class="jumbotron hero jumbotron-fluid">
-            <div class="container">
-              <h1>Census Survey Explorer</h1>
-              <h5>Discover Census Bureau surveys or censuses by name or description or by selecting frequency and / or geography levels.</h5>
-              <input class="quicksearch" placeholder="Search" type="text">
-            </div>
-        </div>
-
-        <div class="container mx-auto">
-            <div id="filterCards" class="mb-3">
-                <div class="btn-group">
-                    <button class="btn btn-primary btn-sm dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                        Geographies
-                        </button>
-                    <div class="dropdown-menu">
-                        <ul>
-                            <li><input type="checkbox" id="National" name="geo" /> National</li> 
-                            <li><input type="checkbox" id="State" name="geo" /> State</li>
-                            <li><input type="checkbox" id="Metro" name="geo" /> Metro Area</li>
-                            <li><input type="checkbox" id="County" name="geo" /> County</li>
-                            <li><input type="checkbox" id="Place" name="geo" /> Place</li>
-                        </ul>
-                        <div class="dropdown-divider"></div>
-                        <a class="dropdown-item" href="#">Learn More</a>
-                    </div>
-                </div>
-                <div class="btn-group">
-                    <button class="btn btn-primary btn-sm dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                        Frequencies
-                    </button>
-                    <div class="dropdown-menu">
-                        <ul>
-                            <li><input type="checkbox" id="Annual" name="freq" /> Annual</li> 
-                            <li><input type="checkbox" id="Quarterly" name="freq" /> Quarterly</li>
-                            <li><input type="checkbox" id="Monthly" name="freq" /> Monthly</li>
-                            <li><input type="checkbox" id="Decennial" name="freq" /> Decennial</li>
-                            <li><input type="checkbox" id="Other" name="freq" /> Other</li>
-                        </ul>
-                        <div class="dropdown-divider"></div>
-                        <a class="dropdown-item" href="#">Learn More</a>
-                    </div>
-                </div>
-                <div class="btn-group">
-                    <button class="btn btn-primary btn-sm dropdown-toggle"  type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                        Subtopics
-                        </button>
-                        <div class="dropdown-menu">
-                            <h6 class="dropdown-header">Demographics</h6>
-                            <ul>
-                                <li><input type="checkbox" id="Age" name="topic" /> Age</li> 
-                                <li><input type="checkbox" id="Ancestry" name="topic" /> Ancestry</li>
-                                <li><input type="checkbox" id="Business" name="topic" /> Business</li>
-                                <li><input type="checkbox" id="BusinessOwner" name="topic" /> Business Owner</li>
-                                <li><input type="checkbox" id="Children" name="topic" /> Children</li>
-                            </ul>
-                            <div class="dropdown-divider"></div>
-                            <a class="dropdown-item" href="#">Learn More</a>
-                        </div>
-                </div>
-
-                <button class="btn btn-outline-secondary btn-sm" onclick="viewChange()">Change View</button>
-                <button class="btn btn-outline-secondary btn-sm" onclick="window.location.reload();">Reset</button>
-            </div>
-            <div id="cards" class="card-columns">
-            </div>
-        </div>
-
-       
-        
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
-        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-        <script src="node_modules/airtable/build/airtable.browser.js"></script>
-        <script src="test.js"></script>
-
-        <script>
-
-            $(document).ready(debounce(function(){
-            $('.quicksearch').on("keyup", function() {
-                var value = $(this).val().toLowerCase();
-                $(".card").filter(function() {
-                $(this).toggle($(this).text().toLowerCase().indexOf(value) > -1)
-                });
-            });
-            }), 200);
-
-            function debounce( fn, threshold ) {
-            var timeout;
-            return function debounced() {
-            if ( timeout ) {
-                clearTimeout( timeout );
-            }
-            function delayed() {
-                fn();
-                timeout = null;
-            }
-            timeout = setTimeout( delayed, threshold || 100 );
-            };
-}
-
-
-
-
-
-            var filter = "";
-
-            var filterCards = function() {
-            if (filter) {
-                $(".card").hide();
-                $(filter).show();
-            } else {
-                $(".card").show();
-            }
-            };
-
-            $("#filterCards input").change(function() {
-            if ($(this).is(":checked")) {
-                filter = filter + "." + this.id;
-            } else {
-                filter = filter.replace("." + this.id, "");
-            }
-            filterCards();
-            });
-       
-     
-        </script>
-        
-    </body>
-</html>
